Extract Name and Address to commons
