+++ src/test/java/seedu/address/logic/parser/ParserUtilTest.java
@@ -14,9 +14,9 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
+import seedu.address.model.commons.Address;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.commons.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
